Task 2 - Spotify2YouTube/requirements.txt
flask
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests